# api/docs/manager/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "manager service API documentation",
        "title": "manager service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/intents": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete one or more schedule intents.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Strategies"
                ],
                "summary": "Delete schedule intents",
                "parameters": [
                    {
                        "description": "Intent IDs to delete",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.DeleteScheduleIntentsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.SuccessResponse-github_com_Gthulhu_api_manager_rest_EmptyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/pods/pids": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all pods running on this node with their associated process IDs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pods"
                ],
                "summary": "Get Pod to PID mappings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_decisionmaker_rest.SuccessResponse-rest_GetPodsPIDsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_decisionmaker_rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Gthulhu_api_decisionmaker_rest.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_Gthulhu_api_decisionmaker_rest.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "github_com_Gthulhu_api_decisionmaker_rest.SuccessResponse-rest_GetPodsPIDsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/rest.GetPodsPIDsResponse"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "github_com_Gthulhu_api_manager_rest.EmptyResponse": {
            "type": "object"
        },
        "github_com_Gthulhu_api_manager_rest.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "github_com_Gthulhu_api_manager_rest.SuccessResponse-github_com_Gthulhu_api_manager_rest_EmptyResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_Gthulhu_api_manager_rest.EmptyResponse"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "rest.DeleteScheduleIntentsRequest": {
            "type": "object",
            "properties": {
                "intentIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "rest.GetPodsPIDsResponse": {
            "type": "object",
            "properties": {
                "node_id": {
                    "type": "string"
                },
                "node_name": {
                    "type": "string"
                },
                "pods": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.PodInfo"
                    }
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "rest.PodInfo": {
            "type": "object",
            "properties": {
                "pod_id": {
                    "type": "string"
                },
                "pod_uid": {
                    "type": "string"
                },
                "processes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.PodProcess"
                    }
                }
            }
        },
...
        "rest.PodProcess": {
            "type": "object",
            "properties": {
                "command": {
                    "type": "string"
                },
                "container_id": {
                    "type": "string"
                },
                "pid": {
                    "type": "integer"
                },
                "ppid": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}